test(env): separate real and manual enviornment tests
- Split test_env.c to two files: one for real shell env and one for
manual
- Add labeled printf statements and separators for each test for better
visibility
- Update instructions in comments to clarify usage
